refactor: Remove legacy modification summary modal
