+++ operate/client/src/App/ProcessInstance/ModificationSummaryModal/index.tsx
@@ -17,8 +17,6 @@ import {
   Modal,
   EmptyCell,
 } from './styled';
-import {processInstanceDetailsStore} from 'modules/stores/processInstanceDetails';
-import {getProcessName} from 'modules/utils/instance';
 import {modificationsStore} from 'modules/stores/modifications';
 import {tracking} from 'modules/tracking';
 import {Button} from '@carbon/react';
@@ -32,6 +30,16 @@ import {
   useModificationsByFlowNode,
   useWillAllFlowNodesBeCanceled,
 } from 'modules/hooks/modifications';
+import {useProcessInstance} from 'modules/queries/processInstance/useProcessInstance';
+import {getProcessDefinitionName} from 'modules/utils/instance';
+
+const OPERATION_DISPLAY_NAME = {
+  ADD_TOKEN: 'Add',
+  CANCEL_TOKEN: 'Cancel',
+  MOVE_TOKEN: 'Move',
+  ADD_VARIABLE: 'Add',
+  EDIT_VARIABLE: 'Edit',
+};
 
 const JSONEditor = lazy(async () => {
   const [{loadMonaco}, {JSONEditor}] = await Promise.all([
@@ -55,21 +63,13 @@ const DiffEditor = lazy(async () => {
   return {default: DiffEditor};
 });
 
-const OPERATION_DISPLAY_NAME = {
-  ADD_TOKEN: 'Add',
-  CANCEL_TOKEN: 'Cancel',
-  MOVE_TOKEN: 'Move',
-  ADD_VARIABLE: 'Add',
-  EDIT_VARIABLE: 'Edit',
-};
-
 const ModificationSummaryModal: React.FC<StateProps> = observer(
   ({open, setOpen}) => {
     const willAllFlowNodesBeCanceled = useWillAllFlowNodesBeCanceled();
     const modificationsByFlowNode = useModificationsByFlowNode();
+    const {data: processInstance} = useProcessInstance();
     const flowNodeModificationsTableRef = useRef<HTMLDivElement>(null);
     const variableModificationsTableRef = useRef<HTMLDivElement>(null);
-    const {processInstance} = processInstanceDetailsStore.state;
 
     useLayoutEffect(() => {
       if (
@@ -102,283 +102,288 @@ const ModificationSummaryModal: React.FC<StateProps> = observer(
       return null;
     }
 
-    const processInstanceId = processInstance.id;
-    const hasParentProcess = processInstance.parentInstanceId !== null;
+    const processInstanceId = processInstance?.processInstanceKey;
+    const hasParentProcess = !!processInstance?.parentProcessInstanceKey;
 
     const areModificationsInvalid =
       willAllFlowNodesBeCanceled && hasParentProcess;
 
     return (
-      <Modal
-        modalHeading="Apply Modifications"
-        size="lg"
-        primaryButtonText="Apply"
-        primaryButtonDisabled={
-          areModificationsInvalid ||
-          modificationsStore.state.modifications.length === 0
-        }
-        secondaryButtonText="Cancel"
-        open={open}
-        onRequestClose={() => setOpen(false)}
-        preventCloseOnClickOutside
-        onRequestSubmit={() => {
-          tracking.track({
-            eventName: 'apply-modifications',
-            addToken: modificationsStore.flowNodeModifications.filter(
-              ({operation}) => operation === 'ADD_TOKEN',
-            ).length,
-            cancelToken: modificationsStore.flowNodeModifications.filter(
-              ({operation}) => operation === 'CANCEL_TOKEN',
-            ).length,
-            moveToken: modificationsStore.flowNodeModifications.filter(
-              ({operation}) => operation === 'MOVE_TOKEN',
-            ).length,
-            addVariable: modificationsStore.variableModifications.filter(
-              ({operation}) => operation === 'ADD_VARIABLE',
-            ).length,
-            editVariable: modificationsStore.variableModifications.filter(
-              ({operation}) => operation === 'EDIT_VARIABLE',
-            ).length,
-            isProcessCanceled: willAllFlowNodesBeCanceled,
-          });
+      processInstanceId && (
+        <Modal
+          modalHeading="Apply Modifications"
+          size="lg"
+          primaryButtonText="Apply"
+          primaryButtonDisabled={
+            areModificationsInvalid ||
+            modificationsStore.state.modifications.length === 0
+          }
+          secondaryButtonText="Cancel"
+          open={open}
+          onRequestClose={() => setOpen(false)}
+          preventCloseOnClickOutside
+          onRequestSubmit={() => {
+            tracking.track({
+              eventName: 'apply-modifications',
+              addToken: modificationsStore.flowNodeModifications.filter(
+                ({operation}) => operation === 'ADD_TOKEN',
+              ).length,
+              cancelToken: modificationsStore.flowNodeModifications.filter(
+                ({operation}) => operation === 'CANCEL_TOKEN',
+              ).length,
+              moveToken: modificationsStore.flowNodeModifications.filter(
+                ({operation}) => operation === 'MOVE_TOKEN',
+              ).length,
+              addVariable: modificationsStore.variableModifications.filter(
+                ({operation}) => operation === 'ADD_VARIABLE',
+              ).length,
+              editVariable: modificationsStore.variableModifications.filter(
+                ({operation}) => operation === 'EDIT_VARIABLE',
+              ).length,
+              isProcessCanceled: willAllFlowNodesBeCanceled,
+            });
 
-          modificationsStore.applyModifications({
-            processInstanceId,
-            onSuccess: () => {
-              tracking.track({eventName: 'modification-successful'});
+            modificationsStore.applyModifications({
+              processInstanceId,
+              onSuccess: () => {
+                tracking.track({eventName: 'modification-successful'});
 
-              notificationsStore.displayNotification({
-                kind: 'success',
-                title: 'Modifications applied',
-                isDismissable: true,
-              });
-            },
-            onError: (statusCode: number) => {
-              tracking.track({eventName: 'modification-failed'});
+                notificationsStore.displayNotification({
+                  kind: 'success',
+                  title: 'Modifications applied',
+                  isDismissable: true,
+                });
+              },
+              onError: (statusCode: number) => {
+                tracking.track({eventName: 'modification-failed'});
 
-              notificationsStore.displayNotification({
-                kind: 'error',
-                title: 'Modification failed',
-                subtitle:
-                  statusCode === 403
-                    ? 'You do not have permission'
-                    : 'Unable to apply modifications, please try again.',
-                isDismissable: true,
-              });
-            },
-          });
-          setOpen(false);
-        }}
-      >
-        <p>
-          {`Planned modifications for Process Instance "${getProcessName(
-            processInstance,
-          )} - ${processInstanceId}". Click "Apply" to proceed.`}
-        </p>
+                notificationsStore.displayNotification({
+                  kind: 'error',
+                  title: 'Modification failed',
+                  subtitle:
+                    statusCode === 403
+                      ? 'You do not have permission'
+                      : 'Unable to apply modifications, please try again.',
+                  isDismissable: true,
+                });
+              },
+            });
+            setOpen(false);
+          }}
+        >
+          <p>
+            {`Planned modifications for Process Instance "${getProcessDefinitionName(
+              processInstance,
+            )} - ${processInstanceId}". Click "Apply" to proceed.`}
+          </p>
 
-        {willAllFlowNodesBeCanceled && !hasParentProcess && <Warning />}
-        {areModificationsInvalid && <Error />}
+          {willAllFlowNodesBeCanceled && !hasParentProcess && <Warning />}
+          {areModificationsInvalid && <Error />}
 
-        <Title>Flow Node Modifications</Title>
-        {modificationsStore.flowNodeModifications.length === 0 ? (
-          <EmptyMessage>No planned flow node modifications</EmptyMessage>
-        ) : (
-          <DataTable
-            ref={flowNodeModificationsTableRef}
-            columnsWithNoContentPadding={['delete']}
-            headers={[
-              {header: ' ', key: 'emptyCell'},
-              {
-                header: 'Operation',
-                key: 'operation',
-                width: '25%',
-              },
-              {
-                header: 'Flow Node',
-                key: 'flowNode',
-                width: '25%',
-              },
-              {
-                header: 'Instance Key',
-                key: 'instanceKey',
-                width: '25%',
-              },
-              {
-                header: 'Affected Tokens',
-                key: 'affectedTokens',
-                width: '20%',
-              },
-              {
-                header: ' ',
-                key: 'delete',
-                width: '5%',
-              },
-            ]}
-            rows={modificationsStore.flowNodeModifications.map(
-              (modification, index) => {
-                const flowNodeId =
-                  modification.operation === 'MOVE_TOKEN'
-                    ? modification.targetFlowNode.id
-                    : modification.flowNode.id;
-                return {
-                  id: index.toString(),
-                  emptyCell: <EmptyCell />,
-                  operation: OPERATION_DISPLAY_NAME[modification.operation],
-                  flowNode: (
-                    <TruncatedValueContainer>
-                      {modification.operation === 'MOVE_TOKEN' ? (
-                        <>
-                          <TruncatedValue $hasMultipleTruncatedValue>
+          <Title>Flow Node Modifications</Title>
+          {modificationsStore.flowNodeModifications.length === 0 ? (
+            <EmptyMessage>No planned flow node modifications</EmptyMessage>
+          ) : (
+            <DataTable
+              ref={flowNodeModificationsTableRef}
+              columnsWithNoContentPadding={['delete']}
+              headers={[
+                {header: ' ', key: 'emptyCell'},
+                {
+                  header: 'Operation',
+                  key: 'operation',
+                  width: '25%',
+                },
+                {
+                  header: 'Flow Node',
+                  key: 'flowNode',
+                  width: '25%',
+                },
+                {
+                  header: 'Instance Key',
+                  key: 'instanceKey',
+                  width: '25%',
+                },
+                {
+                  header: 'Affected Tokens',
+                  key: 'affectedTokens',
+                  width: '20%',
+                },
+                {
+                  header: ' ',
+                  key: 'delete',
+                  width: '5%',
+                },
+              ]}
+              rows={modificationsStore.flowNodeModifications.map(
+                (modification, index) => {
+                  const flowNodeId =
+                    modification.operation === 'MOVE_TOKEN'
+                      ? modification.targetFlowNode.id
+                      : modification.flowNode.id;
+                  return {
+                    id: index.toString(),
+                    emptyCell: <EmptyCell />,
+                    operation: OPERATION_DISPLAY_NAME[modification.operation],
+                    flowNode: (
+                      <TruncatedValueContainer>
+                        {modification.operation === 'MOVE_TOKEN' ? (
+                          <>
+                            <TruncatedValue $hasMultipleTruncatedValue>
+                              {modification.flowNode.name}
+                            </TruncatedValue>
+                            &nbsp;→&nbsp;
+                            <TruncatedValue $hasMultipleTruncatedValue>
+                              {modification.targetFlowNode.name}
+                            </TruncatedValue>
+                          </>
+                        ) : (
+                          <TruncatedValue>
                             {modification.flowNode.name}
                           </TruncatedValue>
-                          &nbsp;→&nbsp;
-                          <TruncatedValue $hasMultipleTruncatedValue>
-                            {modification.targetFlowNode.name}
+                        )}
+                      </TruncatedValueContainer>
+                    ),
+                    instanceKey: (
+                      <TruncatedValueContainer>
+                        {modification.operation !== 'ADD_TOKEN' &&
+                        modification.flowNodeInstanceKey !== undefined ? (
+                          <TruncatedValue>
+                            {modification.flowNodeInstanceKey}
                           </TruncatedValue>
-                        </>
+                        ) : (
+                          <>--</>
+                        )}
+                      </TruncatedValueContainer>
+                    ),
+                    affectedTokens: (
+                      <span data-testid="affected-token-count">
+                        {modification.operation === 'CANCEL_TOKEN'
+                          ? modification.visibleAffectedTokenCount +
+                            (modificationsByFlowNode[flowNodeId]
+                              ?.cancelledChildTokens ?? 0)
+                          : modification.affectedTokenCount}
+                      </span>
+                    ),
+                    delete: (
+                      <Button
+                        kind="danger--ghost"
+                        title="Delete flow node modification"
+                        aria-label="Delete flow node modification"
+                        size="sm"
+                        onClick={() => {
+                          modificationsStore.removeFlowNodeModification(
+                            modification,
+                          );
+                        }}
+                      >
+                        Delete
+                      </Button>
+                    ),
+                  };
+                },
+              )}
+            />
+          )}
+          <Title>Variable Modifications</Title>
+          {modificationsStore.variableModifications.length === 0 ? (
+            <EmptyMessage>No planned variable modifications</EmptyMessage>
+          ) : (
+            <DataTable
+              ref={variableModificationsTableRef}
+              columnsWithNoContentPadding={['delete']}
+              isExpandable
+              expandableRowTitle="View variable changes by expanding rows; editing disabled in read-only editor."
+              expandedContents={modificationsStore.variableModifications.reduce(
+                (
+                  accumulator,
+                  {id, scopeId, operation, oldValue, newValue},
+                ) => ({
+                  ...accumulator,
+                  [`${scopeId}/${id}`]: (
+                    <Suspense>
+                      {operation === 'ADD_VARIABLE' ? (
+                        <JSONEditor
+                          value={beautifyJSON(newValue)}
+                          readOnly
+                          height="10vh"
+                          width="95%"
+                        />
                       ) : (
-                        <TruncatedValue>
-                          {modification.flowNode.name}
-                        </TruncatedValue>
+                        <DiffEditor
+                          modifiedValue={beautifyJSON(newValue)}
+                          originalValue={beautifyJSON(oldValue ?? '')}
+                          height="10vh"
+                          width="95%"
+                        />
                       )}
-                    </TruncatedValueContainer>
+                    </Suspense>
                   ),
-                  instanceKey: (
-                    <TruncatedValueContainer>
-                      {modification.operation !== 'ADD_TOKEN' &&
-                      modification.flowNodeInstanceKey !== undefined ? (
-                        <TruncatedValue>
-                          {modification.flowNodeInstanceKey}
-                        </TruncatedValue>
-                      ) : (
-                        <>--</>
-                      )}
-                    </TruncatedValueContainer>
-                  ),
-                  affectedTokens: (
-                    <span data-testid="affected-token-count">
-                      {modification.operation === 'CANCEL_TOKEN'
-                        ? modification.visibleAffectedTokenCount +
-                          (modificationsByFlowNode[flowNodeId]
-                            ?.cancelledChildTokens ?? 0)
-                        : modification.affectedTokenCount}
-                    </span>
-                  ),
-                  delete: (
-                    <Button
-                      kind="danger--ghost"
-                      title="Delete flow node modification"
-                      aria-label="Delete flow node modification"
-                      size="sm"
-                      onClick={() => {
-                        modificationsStore.removeFlowNodeModification(
-                          modification,
-                        );
-                      }}
-                    >
-                      Delete
-                    </Button>
-                  ),
-                };
-              },
-            )}
-          />
-        )}
-        <Title>Variable Modifications</Title>
-        {modificationsStore.variableModifications.length === 0 ? (
-          <EmptyMessage>No planned variable modifications</EmptyMessage>
-        ) : (
-          <DataTable
-            ref={variableModificationsTableRef}
-            columnsWithNoContentPadding={['delete']}
-            isExpandable
-            expandableRowTitle="View variable changes by expanding rows; editing disabled in read-only editor."
-            expandedContents={modificationsStore.variableModifications.reduce(
-              (accumulator, {id, scopeId, operation, oldValue, newValue}) => ({
-                ...accumulator,
-                [`${scopeId}/${id}`]: (
-                  <Suspense>
-                    {operation === 'ADD_VARIABLE' ? (
-                      <JSONEditor
-                        value={beautifyJSON(newValue)}
-                        readOnly
-                        height="10vh"
-                        width="95%"
-                      />
-                    ) : (
-                      <DiffEditor
-                        modifiedValue={beautifyJSON(newValue)}
-                        originalValue={beautifyJSON(oldValue ?? '')}
-                        height="10vh"
-                        width="95%"
-                      />
-                    )}
-                  </Suspense>
-                ),
-              }),
-              {},
-            )}
-            headers={[
-              {
-                header: 'Operation',
-                key: 'operation',
-                width: '25%',
-              },
-              {
-                header: 'Scope',
-                key: 'scope',
-                width: '25%',
-              },
-              {
-                header: 'Name / Value',
-                key: 'nameValue',
-                width: '25%',
-              },
-              {
-                header: ' ',
-                key: 'emptyCell',
-                width: '20%',
-              },
-              {
-                header: ' ',
-                key: 'delete',
-                width: '5%',
-              },
-            ]}
-            rows={modificationsStore.variableModifications.map(
-              ({operation, flowNodeName, name, newValue, scopeId, id}) => {
-                return {
-                  id: `${scopeId}/${id}`,
-                  operation: OPERATION_DISPLAY_NAME[operation],
-                  scope: <TruncatedValue>{flowNodeName}</TruncatedValue>,
-                  nameValue: (
-                    <VariableModification name={name} newValue={newValue} />
-                  ),
-                  emptyCell: '',
-                  delete: (
-                    <Button
-                      kind="danger--ghost"
-                      title="Delete variable modification"
-                      aria-label="Delete variable modification"
-                      size="sm"
-                      onClick={() => {
-                        modificationsStore.removeVariableModification(
-                          scopeId,
-                          id,
-                          operation,
-                          'summaryModal',
-                        );
-                      }}
-                    >
-                      Delete
-                    </Button>
-                  ),
-                };
-              },
-            )}
-          />
-        )}
-      </Modal>
+                }),
+                {},
+              )}
+              headers={[
+                {
+                  header: 'Operation',
+                  key: 'operation',
+                  width: '25%',
+                },
+                {
+                  header: 'Scope',
+                  key: 'scope',
+                  width: '25%',
+                },
+                {
+                  header: 'Name / Value',
+                  key: 'nameValue',
+                  width: '25%',
+                },
+                {
+                  header: ' ',
+                  key: 'emptyCell',
+                  width: '20%',
+                },
+                {
+                  header: ' ',
+                  key: 'delete',
+                  width: '5%',
+                },
+              ]}
+              rows={modificationsStore.variableModifications.map(
+                ({operation, flowNodeName, name, newValue, scopeId, id}) => {
+                  return {
+                    id: `${scopeId}/${id}`,
+                    operation: OPERATION_DISPLAY_NAME[operation],
+                    scope: <TruncatedValue>{flowNodeName}</TruncatedValue>,
+                    nameValue: (
+                      <VariableModification name={name} newValue={newValue} />
+                    ),
+                    emptyCell: '',
+                    delete: (
+                      <Button
+                        kind="danger--ghost"
+                        title="Delete variable modification"
+                        aria-label="Delete variable modification"
+                        size="sm"
+                        onClick={() => {
+                          modificationsStore.removeVariableModification(
+                            scopeId,
+                            id,
+                            operation,
+                            'summaryModal',
+                          );
+                        }}
+                      >
+                        Delete
+                      </Button>
+                    ),
+                  };
+                },
+              )}
+            />
+          )}
+        </Modal>
+      )
     );
   },
 );
