+++ operate/client/src/App/ProcessInstance/index.tsx
@@ -26,7 +26,7 @@ import {FlowNodeInstanceLog} from './FlowNodeInstanceLog/v2';
 import {Button, Modal} from '@carbon/react';
 import {tracking} from 'modules/tracking';
 import {ModalStateManager} from 'modules/components/ModalStateManager';
-import {ModificationSummaryModal} from './ModificationSummaryModal/v2';
+import {ModificationSummaryModal} from './ModificationSummaryModal';
 import {useCallbackPrompt} from 'modules/hooks/useCallbackPrompt';
 import {LastModification} from './LastModification';
 import {VariablePanel} from './BottomPanel/VariablePanel';
